Magerun2 fails outside of Magento 2 (#38)
With this module enabled, the running of `magerun2` outside a Magento root, throws a warning, because this module already calls upon the M2 code in its constructor. This PR refactors things so that M2 classes are only called upon once we know that the `magerun2` command is executed within the M2 root.