Blacklist Bss_MultiWishlist (CWE-79) (#64)
* Blacklist Bss_MultiWishlist (CWE-79)

More information can be found at https://xn--gran-8qa.fi/bss-multiwishlist-xss-injection/ after given 3 days to vendor has passed. I have provided them also a fix for this issue, so it's just a matter of releasing a patching version. Fixing version will be added, when known is available.

I'm requesting this to be added on the list.

Reference for BSS: 17573.

* Add fixed_in for BSS_MultiWishlist